+++ lib/presentation/home/views/widgets/details/book_details_view_body.dart
@@ -2,7 +2,7 @@ import 'package:bookly/presentation/home/views/widgets/common/book_rating.dart';
 import 'package:bookly/presentation/home/views/widgets/details/book_details_image.dart';
 import 'package:bookly/presentation/home/views/widgets/details/book_price_preview_widget.dart';
 import 'package:bookly/presentation/home/views/widgets/details/custom_book_details_appbar.dart';
-import 'package:bookly/presentation/home/views/widgets/details/see_also_books_listview.dart';
+import 'package:bookly/presentation/home/views/widgets/details/see_also_books_section.dart';
 
 import 'package:bookly/presentation/resources/colors.dart';
 import 'package:bookly/presentation/resources/strings.dart';
@@ -38,16 +38,7 @@ class BookDetailsViewBody extends StatelessWidget {
         const SizedBox(height: AppSizes.s30),
         const BookPriceAndPreview(),
         const SizedBox(height: AppSizes.s20),
-        Container(
-          alignment: AlignmentDirectional.centerStart,
-          padding: const EdgeInsets.only(left: AppPadding.p30),
-          child: const Text(
-            AppStrings.youCanAlsoLike,
-            style: AppStyles.textStyle14,
-          ),
-        ),
-        const SizedBox(height: AppSizes.s5),
-        const SeeAlsoBooksListView(),
+        const SeeAlsoBooksSection(),
         const SizedBox(height: AppSizes.s20),
       ],
     );
